+++ src/event/index.ts
@@ -1,4 +1,4 @@
-import { InjectionKey, inject, onBeforeUnmount } from 'vue';
+import { onBeforeUnmount } from 'vue';
 
 export type Handler<T = any> = (event: T) => any;
 export interface EventsStore {
@@ -6,13 +6,11 @@ export interface EventsStore {
   handler: Handler;
 }
 export interface WowerlayContext {
-  onRecalculate: (handler: Handler<WheelEvent | Event>) => void;
+  onRecalculate: (handler: Handler<WheelEvent | Event | undefined>) => void;
   onWindowClick: (handler: Handler<MouseEvent>) => void;
   calculateAll: (event: WheelEvent | Event) => void;
   clickAll: (event: PointerEvent) => void;
 }
-
-export const injectionKey: InjectionKey<WowerlayContext> = Symbol();
 
 export const createEvent = (store: EventsStore[], handler: Handler) => {
   const id = Symbol();
@@ -32,7 +30,3 @@ export const runEvents = (store: EventsStore[], e?: any) => {
 };
 
 export const createEventStore = () => [] as EventsStore[];
-
-export const useWowerlayContext = () => {
-  return inject(injectionKey)!;
-};